Added another call to the function "nextSliderImg()" before starts setInterval.
Because setInterval waits 3 seconds for every call, but the first element has to be
without the className "hidden" from the beginning, when the DOMcontent is Loaded.
Also added padding in the width of the container logos, because You could see
the square corners of the images when the screen width is less than 770px.