Changed the way of create the modal.
I remove all the parameters of the function createModal and leave only one
the "e" of event, so titleEvent reads the h3 of card and filters
the array of all elements of the json. With this We can trat each event
seàrately and put conditionals inside the createModal function without
the need of external parameters.