Added iframe with API Google Embed Map
URLGOOGLE/MAP_MODE?key=API_KEY&PARAMETERS(q=${eventSelect.site})
API_KEY does not need to be hidden because It is not a secret key.
MAP_MODE is the type of the map. In our case, We only need the display of the place. So I write Place
PARAMETERS the only required parameter with the MAP_MODE= place is q. Which refers to the place name ,
address, plus Code or place ID.